+++ SourceryTests/Parsing/FileParserSpec.swift
@@ -484,18 +484,18 @@ class FileParserSpec: QuickSpec {
                                           ]))
                     }
 
-                    it("does not include default values for asssociated types as part of its name") {
+                    it("extracts default values for asssociated values") {
                         expect(parse("enum Foo { case optionA(Int = 1, named: Float = 42.0, _: Bool = false); case optionB(Bool = true) }"))
                         .to(equal([
                             Enum(name: "Foo", accessLevel: .internal, isExtension: false, inheritedTypes: [], cases:
                                 [
                                     EnumCase(name: "optionA", associatedValues: [
-                                        AssociatedValue(localName: nil, externalName: "0", typeName: TypeName("Int")),
-                                        AssociatedValue(localName: "named", externalName: "named", typeName: TypeName("Float")),
-                                        AssociatedValue(localName: nil, externalName: "2", typeName: TypeName("Bool"))
+                                        AssociatedValue(localName: nil, externalName: "0", typeName: TypeName("Int"), defaultValue: "1"),
+                                        AssociatedValue(localName: "named", externalName: "named", typeName: TypeName("Float"), defaultValue: "42.0"),
+                                        AssociatedValue(localName: nil, externalName: "2", typeName: TypeName("Bool"), defaultValue: "false")
                                         ]),
                                     EnumCase(name: "optionB", associatedValues: [
-                                    AssociatedValue(localName: nil, externalName: nil, typeName: TypeName("Bool"))
+                                    AssociatedValue(localName: nil, externalName: nil, typeName: TypeName("Bool"), defaultValue: "true")
                                     ])
                                 ])
                         ]))