+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Added `optional` filter for variables
 - Added `json` filter to output raw JSON objects 
+- Added `.defaultValue` to `AssociatedValue`
 
 ### Internal Changes